// packages/member-services-hub/src/components/Headerbar/types.ts
export interface HeaderbarProps {
  firstname: string;
  surname: string;
  email: string;
  button?: boolean;
  rightChevron?: boolean;
  leftChevron?: boolean;
  notifications?: boolean;
  messages?: boolean;
  calender?: boolean;
  welcome?: boolean;
  welcomeHeader?: string;
  welcomeText?: string;
  search?: boolean;
  profilePicture?: string;
  buttonText?: string;
}